# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "César Core",
    "version": "1.1.0"
  },
  "paths": {
...
    "/v1/search": {
      "post": {
        "tags": [
          "search"
        ],
        "summary": "Search Web",
        "operationId": "search_web_v1_search_post",
        "security": [
          {
            "ApplicationBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-Service",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Service"
            }
          },
          {
            "name": "X-Purpose",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Purpose"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequestPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SecurityErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SecurityErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/fetch": {
      "post": {
        "tags": [
          "fetch"
        ],
        "summary": "Fetch Url",
        "operationId": "fetch_url_v1_fetch_post",
        "security": [
          {
            "ApplicationBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-Service",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Service"
            }
          },
          {
            "name": "X-Purpose",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Purpose"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FetchRequestPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FetchResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FetchErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FetchErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SecurityErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SecurityErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FetchErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FetchErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "FetchErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/FetchErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "FetchErrorResponse",
        "description": "Envelope público de erro do Central Web Fetch/Enrichment Gateway."
      },
      "FetchRequestPayload": {
        "properties": {
          "requirements": {
            "$ref": "#/components/schemas/Requirements"
          },
          "url": {
            "type": "string",
            "maxLength": 2048,
            "minLength": 1,
            "title": "Url"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "requirements",
          "url"
        ],
        "title": "FetchRequestPayload",
        "description": "DTO HTTP público de uma requisição de enriquecimento de URL.\n\nNão recebe ``application_id`` nem qualquer outro dado de identidade no\ncorpo: quem está chamando é sempre resolvido pelo Core, nunca declarado\npelo cliente dentro do payload. Não recebe nome de provider nem opção\nespecífica de provider (formato, profundidade, seletor): esta é a\nfronteira provider-neutral -- o adapter concreto decide isso."
      },
      "FetchResponse": {
        "properties": {
          "request_id": {
            "type": "string",
            "minLength": 1,
            "title": "Request Id"
          },
          "correlation_id": {
            "type": "string",
            "minLength": 1,
            "title": "Correlation Id"
          },
          "provider_gateway": {
            "type": "string",
            "minLength": 1,
            "title": "Provider Gateway"
          },
          "provider": {
            "type": "string",
            "minLength": 1,
            "title": "Provider"
          },
          "url": {
            "type": "string",
            "minLength": 1,
            "title": "Url"
          },
          "fetched": {
            "type": "boolean",
            "title": "Fetched"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "truncated": {
            "type": "boolean",
            "title": "Truncated",
            "default": false
          },
          "usage": {
            "$ref": "#/components/schemas/FetchUsage"
          },
          "latency_ms": {
            "type": "number",
            "minimum": 0.0,
            "title": "Latency Ms"
          },
          "upstream_request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Upstream Request Id"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "correlation_id",
          "provider_gateway",
          "provider",
          "url",
          "fetched",
          "latency_ms"
        ],
        "title": "FetchResponse",
        "description": "Forma neutra de uma resposta de enriquecimento devolvida pelo Core.\n\n``fetched=False`` é sucesso de negócio, não erro: a URL foi alcançada\npelo provider, mas a origem específica não devolveu conteúdo aproveitável\n(bloqueio/paywall/página vazia). O chamador nunca recebe uma exceção\nnesse caso -- mesma semântica que ``FirecrawlScrapeProvider.scrape_basic``\njá tinha no GG Oferta antes desta capability existir no Core."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Requirements": {
        "properties": {
          "service_class": {
            "$ref": "#/components/schemas/ServiceClass"
          },
          "cost_policy": {
            "$ref": "#/components/schemas/CostPolicy"
          }
        },
        "type": "object",
        "required": [
          "service_class",
          "cost_policy"
        ],
        "title": "Requirements",
        "description": "O que uma execução exige em termos de qualidade/custo.\n\nResponsabilidade final (ver ADR 0009): ``ApplicationContext`` é\nquem está chamando + por quê + identidade/tracing da chamada;\n``Requirements`` é quais capacidades/qualidade/custo a execução\nexige. Por isso não há campo ``service`` aqui -- ele já existe em\n``ApplicationContext`` (o serviço/processo chamador) e um segundo\ncampo de mesmo nome, com significado diferente (qual gateway:\nai/search), era exatamente a duplicação que a TASK-118A pediu para\nremover. Qual gateway (AI vs Search) já é dado pelo tipo do request\n(``AIRequest`` vs ``SearchRequest``), não precisa ser repetido aqui.\n\n``structured_output``, ``reasoning``, ``vision``, ``tool_calling``\n(capacidades específicas) não foram adicionados nesta correção:\nficam para quando uma TASK futura precisar deles de fato."
      },
      "SearchErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "minLength": 1,
            "title": "Code"
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "title": "Message"
          },
          "request_id": {
            "type": "string",
            "minLength": 1,
            "title": "Request Id"
          },
          "correlation_id": {
            "type": "string",
            "minLength": 1,
            "title": "Correlation Id"
          },
          "upstream_request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Upstream Request Id"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "request_id",
          "correlation_id"
        ],
        "title": "SearchErrorDetail",
        "description": "Detalhe de erro estável do endpoint Search."
      },
      "SearchErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/SearchErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "SearchErrorResponse",
        "description": "Envelope público de erro do Central Web Search Gateway."
      },
      "SearchRequestPayload": {
        "properties": {
          "requirements": {
            "$ref": "#/components/schemas/Requirements"
          },
          "query": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Query"
          },
          "max_results": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Max Results",
            "description": "Limite obrigatório dos resultados expostos. Encaminhado ao upstream; não garante limite de aquisição externa quando o provider não o suporta.",
            "default": 5
          }
        },
        "type": "object",
        "required": [
          "requirements",
          "query"
        ],
        "title": "SearchRequestPayload",
        "description": "DTO HTTP público de uma requisição de busca.\n\nNão recebe ``application_id`` nem qualquer outro dado de identidade\nno corpo: quem está chamando é sempre resolvido pelo Core, nunca\ndeclarado pelo cliente dentro do payload."
      },
      "SearchResponse": {
        "properties": {
          "request_id": {
            "type": "string",
            "minLength": 1,
            "title": "Request Id"
          },
          "correlation_id": {
            "type": "string",
            "minLength": 1,
            "title": "Correlation Id"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/SearchResult"
            },
            "type": "array",
            "title": "Results"
          },
          "provider_gateway": {
            "type": "string",
            "minLength": 1,
            "title": "Provider Gateway"
          },
          "provider": {
            "type": "string",
            "minLength": 1,
            "title": "Provider"
          },
          "usage": {
            "$ref": "#/components/schemas/SearchUsage"
          },
          "latency_ms": {
            "type": "number",
            "minimum": 0.0,
            "title": "Latency Ms"
          },
          "total_results_available": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Results Available"
          },
          "cached": {
            "type": "boolean",
            "title": "Cached",
            "default": false
          },
          "fallback_used": {
            "type": "boolean",
            "title": "Fallback Used",
            "default": false
          },
          "upstream_request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Upstream Request Id"
          },
          "upstream_issues": {
            "items": {
              "$ref": "#/components/schemas/SearchUpstreamIssue"
            },
            "type": "array",
            "title": "Upstream Issues"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "correlation_id",
          "provider_gateway",
          "provider",
          "usage",
          "latency_ms"
        ],
        "title": "SearchResponse",
        "description": "Forma neutra de uma resposta de busca devolvida pelo César Core."
      },
      "SearchResult": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "snippet": {
            "type": "string",
            "title": "Snippet",
            "default": ""
          },
          "position": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          },
          "score": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          }
        },
        "type": "object",
        "required": [
          "title",
          "url"
        ],
        "title": "SearchResult",
        "description": "Um resultado individual de busca."
      },
      "SearchUpstreamIssue": {
        "properties": {
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "provider",
          "code",
          "message"
        ],
        "title": "SearchUpstreamIssue",
        "description": "Falha parcial reportada pelo upstream sem invalidar a resposta."
      },
      "SearchUsage": {
        "properties": {
          "queries_used": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Queries Used"
          },
          "search_cost_usd": {
            "type": "number",
            "minimum": 0.0,
            "title": "Search Cost Usd"
          },
          "llm_tokens": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Tokens"
          }
        },
        "type": "object",
        "required": [
          "queries_used",
          "search_cost_usd"
        ],
        "title": "SearchUsage",
        "description": "Uso normalizado da busca informado pelo gateway upstream."
      },
      "SecurityErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "minLength": 1,
            "title": "Code"
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "title": "Message"
          },
          "request_id": {
            "type": "string",
            "minLength": 1,
            "title": "Request Id"
          },
          "correlation_id": {
            "type": "string",
            "minLength": 1,
            "title": "Correlation Id"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "request_id",
          "correlation_id"
        ],
        "title": "SecurityErrorDetail"
      },
      "SecurityErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/SecurityErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "SecurityErrorResponse"
      },
      "ServiceClass": {
        "type": "string",
        "enum": [
          "economy",
          "standard",
          "quality"
        ],
        "title": "ServiceClass",
        "description": "Nível de custo/qualidade desejado para atender a uma requisição."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "ApplicationBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}